pyproject.toml
[tool.poetry]
name = "f1-graph"
version = "0.1.0"
description = ""
authors = ["Hector Casanova <hec1802@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
py2neo = "^2021.2.4"
pandas = "^2.2.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
